extract namespaced_data
